Moved JavaXml
